ao_coreaudio_exclusive: always reload on stream format change
This is a simplification.

I'm not sure of the significance of this code, or when it even happens.
Assume that it happens rarely, and in particular assume it doesn't
change between two passthrough formats all the time.

It is possible that the behavior of this changed in commit 2f5e31cf
anyway.

WIP: actually, this happens on plug events, even if the format doesn't
change, so the format should be compared here.